do pass groups individually to auth backend
As discussed in https://github.com/splitbrain/dokuwiki/issues/3028, what
can be passed as a $filter to retrieveUsers() is ill defined right now.
Assuming that we can pass a | to fetch users from two different groups
is bound to fail with various backends.

This patch will ask for members of all given groups separately and merge
them in code. By passing a simple string (not a pattern) this should be
the most compatible implementation for the current situation.

It might be a good idea to allow an array to be passed in the grps
filter (or even in all filters?) to signal that you want to match
multiple groups. This should be discussed in the ticket above.